Created Games class
Game class initialization enabled

The Game class allows instantiating the Swing interface and managing the game flow. Added mouse listener to handle mouse clicks.
The initializeGame() method was created to start the game depending on the number of players.
Created components that display player information and game status when the game is started.